Add new RELGCD pass to remove global lists of empty constructor or destructor
Remove empty list of global constructors or destructors (RELGCD).

The global dead-code elimination remove the useless constructors and
destructors from the code and also from "@llvm.global_ctors" and
"@llvm.global_dtors".

But when these arrays are empty they remain in the code and that chokes some
SPIR consumers such as Xilinx xocc, as it is not legal SPIR:
\code
@llvm.global_ctors = appending global [0 x { i32, void ()*, i8* }] zeroinitializer
\endcode
So this pass removes these empty "@llvm.global_ctors" and
"@llvm.global_dtors"