Modify LoopIdiomRecognize.cpp to skip loop idiom optimization for kernel and functions called in the kernel
This branch is aim to solve problem trigger in this example:
https://gitenterprise.xilinx.com/linyay/triSYCL/commit/2a0dbdb485293c32d1ce380a68d35ea897269b06#diff-4beb17f5a3dbac38d7cb72c001c1b27d

Since argument promotion can only be done with GEP instructions that
only used by loads. And if there are llvm.memcpy instruction, there
will always a bitcast instruction before to cast the pointer, so force
kernel and functions called in the kernel in per_kernel_caller.ll to
skip loop idiom opt pass.